Refactor maybe-rayon (#217)
* refactor maybe-rayon

* fmt reorder import

* rm comment
