+++ matrix/src/util.rs
@@ -1,4 +1,4 @@
-use p3_maybe_rayon::{MaybeIntoParIter, ParallelIterator};
+use p3_maybe_rayon::prelude::*;
 use p3_util::{log2_strict_usize, reverse_bits_len};
 
 use crate::dense::RowMajorMatrix;
