+++ maybe-rayon/Cargo.toml
@@ -6,7 +6,6 @@ version = "0.1.0"
 edition = "2021"
 
 [features]
-default = ["parallel"]
 parallel = ["rayon"]
 
 [dependencies]
