+++ merkle-tree/src/merkle_tree.rs
@@ -7,7 +7,7 @@ use itertools::Itertools;
 use p3_field::{AbstractField, Field, PackedField};
 use p3_matrix::dense::RowMajorMatrix;
 use p3_matrix::{Matrix, MatrixRowSlices};
-use p3_maybe_rayon::{IndexedParallelIterator, MaybeParChunksMut, ParallelIterator};
+use p3_maybe_rayon::prelude::*;
 use p3_symmetric::{CryptographicHasher, PseudoCompressionFunction};
 use tracing::instrument;
